rename serialize/deserialize & lint fix
